pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lasermx"
version = "0.1.0"
description = "LaserMX - Software latino de control de corte/grabado láser compatible con GRBL (macOS)"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "pyserial>=3.5",
    "PySide6>=6.6",
    "svgpathtools>=1.6.1",
    "ezdxf>=1.3",
    "shapely>=2.0",
    "typer>=0.12",
    "rich>=13.7"
]

[project.scripts]
lasermx = "lasermx.cli:main"
lasermx-gui = "lasermx.gui.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["lasermx*"]
exclude = ["tests*", "examples*"]